requirements.txt
pytest>=8,<9
pytest-playwright>=0.7,<1
pytest-base-url>=2.1,<3
playwright>=1.58,<2
allure-pytest>=2.13,<3
pytest-bdd>=8,<9
